Move queries into a proper folder
