+++ app/graphql/resolvers/plans_resolver.rb
@@ -17,7 +17,7 @@ module Resolvers
     def resolve(ids: nil, page: nil, limit: nil, search_term: nil)
       validate_organization!
 
-      query = ::Queries::PlanQuery.new(organization: current_organization)
+      query = PlanQuery.new(organization: current_organization)
       result = query.call(
         search_term:,
         page:,
